flashmem test
