+++ CMakeLists.txt
@@ -26,6 +26,7 @@ add_subdirectory(test/Unity Unity_build)
 add_executable(test_${PROJECT_NAME}
   test/test_main.c
   test/test_some_module.c
+  test/test_flashmem.c
 )
 
 pico_set_linker_script(${PROJECT_NAME} ${CMAKE_SOURCE_DIR}/memmap_custom.ld)
