+++ src/flashmem/flashmem.h
@@ -1,7 +1,16 @@
 #ifndef FLASHMEM_H
 #define FLASHMEM_H
 
-int write(char* key, char* val);
-int read(char* key, char* out_val);
+enum _flashmem_err 
+{
+  FM_ERR_OK = 0,
+  FM_ERR_FLASHMEM_SIZE = 1,
+  FM_ERR_INVALID_KEY = 2,
+  FM_ERR_CORRUPT = 3,
+};
+typedef enum _flashmem_err flashmem_err_t;
+
+flashmem_err_t write(const char* key, const char* val);
+flashmem_err_t read(const char* key, char* out_val);
 
 #endif
