+++ test/test_main.c
@@ -4,6 +4,7 @@
 #include "pico/stdlib.h"
 
 extern void test_my_module_function(void);
+extern void test_flashmem(void);
 
 int main(void) {
     stdio_init_all(); // Initialize UART
