+++ test/test_some_module.c
@@ -1,4 +1,3 @@
-// test/test_some_module.c
 #include "unity.h"
 
 void setUp(void) {
@@ -13,10 +12,3 @@ void test_my_module_function(void) {
     // Example test case
     TEST_ASSERT_EQUAL_INT(5, 5);  // Asserts that my_module_function(2, 3) returns 5
 }
-
-/* int main(void) { */
-/*     UNITY_BEGIN(); */
-/*     RUN_TEST(test_my_module_function);  // Add all your test functions here */
-/*     return UNITY_END(); */
-/* } */
-
